app/requirements.txt
Django==2.2.*
djangorestframework==3.10.2
python-dateutil==2.8.*
requests==2.23.0